Change variable delimiter
Since the `__' is used quite a lot in, for example, Python code and
files it proved to be a sub-optimal as a delimiter. Now, by default,
`$!' is used, which should appear somewhat less (I hope) and it has
been made customizable. Look out when customizing, it is placed in a
regular expression without any escaping.